MPVActivity: only exit at IDLE event after START_FILE is received
We were always supposed to get IDLE during start-up, but a bug in
mpv was actually making it not happen. This in turn made us think
our code was working properly. Alas, in reality, it wasn't.

The bug on mpv's side is now fixed and would cause early exits at
start-up if the code was not changed. Thus we add a flag to signal
if input has been started at least once in the current session and
only exit with an IDLE state after that flag has been set.