MPVActivity: Add pref to extend video in the cutout area
Many newer Android devices feature a notch or hole-punch camera, and by
default, Android will size landscape windows so that they don't extend
into the cutout area. For video playback, it is common to support
displaying the video over the entire screen, regardless of the cutout,
so let's add that option.

I've defaulted this to on as most people want to use all their screen.