MPVActivity: account for cutout inset at the top
I want to take a short moment here to say how useless the APIs are.
You might think that with WindowInsetsCompat.Type.statusBars() you could
retrieve the amount of space Android forcibly reserves for the status bar
and then subtract it from the existing insets, thereby creating a
portable solution that should reliably work in a variety of situations.
Wrong, it always returns (0, 0, 0, 0).
So now we we have this worse hack that probably malfunctions if you hold
your phone upside down.